transparent chromatic stories set
