+++ packages/react-components/src/select/tests/chromatic/transparent/transparent.chroma.jsx
@@ -2,8 +2,22 @@ import { CalendarIcon, MagnifierIcon } from "@react-components/icons";
 import { Select } from "@react-components/select";
 import { createChromaticSection, paramsBuilder, storiesOfBuilder } from "@utils";
 
+const MALE = { key: "Male", text: "Male", value: "Male" };
+const FEMALE = { key: "Female", text: "Female", value: "Female" };
+const GENDERS = [MALE, FEMALE];
+
+function createSelect({ options = GENDERS, ...otherProps } = {}) {
+    return <Select
+        placeholder="Gender"
+        search
+        transparent
+        options={options}
+        {...otherProps}
+    />;
+}
+
 function stories(segment) {
-    return storiesOfBuilder(module, createChromaticSection("Select/searchable"))
+    return storiesOfBuilder(module, createChromaticSection("Select/transparent"))
         .segment(segment)
         .parameters(paramsBuilder()
             .chromaticDelay(100)
@@ -12,89 +26,151 @@ function stories(segment) {
         .build();
 }
 
-const MALE = { key: "Male", text: "Male", value: "Male" };
-const FEMALE = { key: "Female", text: "Female", value: "Female" };
-const GENDERS = [MALE, FEMALE];
-
-function createSearchableSelect({ options = GENDERS, ...otherProps } = {}) {
-    return <Select
-        placeholder="Gender"
-        search
-        options={options}
-        {...otherProps}
-    />;
-}
-
 stories()
     .add("default", () =>
         <div className="flex">
-            {createSearchableSelect({
+            {createSelect({
                 className: "mr5"
             })}
-            {createSearchableSelect({
+            {createSelect({
                 defaultOpen: true
             })}
         </div>
     )
-    .add("size", () =>
+    .add("selected value", () =>
+        <div className="flex">
+            {createSelect({
+                defaultValue: "Female",
+                className: "mr5"
+            })}
+            {createSelect({
+                defaultValue: "Female",
+                defaultOpen: true
+            })}
+        </div>
+    )
+    .add("size",
+         () =>
+             <div className="flex flex-column">
+                 <div className="flex" style={{ marginBottom: "150px" }}>
+                     {createSelect({
+                         size: "small",
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         size: "small",
+                         defaultValue: "Female",
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         size: "small",
+                         defaultOpen: true
+                     })}
+                 </div>
+                 <div className="flex" style={{ marginBottom: "150px" }}>
+                     {createSelect({
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         defaultValue: "Female",
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         defaultOpen: true
+                     })}
+                 </div>
+                 <div className="flex">
+                     {createSelect({
+                         size: "large",
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         size: "large",
+                         defaultValue: "Female",
+                         className: "mr5"
+                     })}
+                     {createSelect({
+                         size: "large",
+                         defaultOpen: true
+                     })}
+                 </div>
+             </div>
+    )
+    .add("fluid", () =>
         <div className="flex flex-column">
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
-                    className: "mr5",
-                    size: "small"
-                })}
-                {createSearchableSelect({
-                    defaultValue: "Male",
-                    className: "mr5",
-                    size: "small"
-                })}
-                {createSearchableSelect({
-                    size: "small",
-                    defaultOpen: true
-                })}
-            </div>
-            <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
+                    fluid: true,
                     className: "mr5"
                 })}
-                {createSearchableSelect({
-                    defaultValue: "Male",
-                    className: "mr5"
-                })}
-                {createSearchableSelect({
+                {createSelect({
+                    fluid: true,
                     defaultOpen: true
                 })}
             </div>
             <div className="flex">
-                {createSearchableSelect({
-                    className: "mr5",
+                <div className="w-30 mr5">
+                    {createSelect({
+                        fluid: true
+                    })}
+                </div>
+                <div className="w-30">
+                    {createSelect({
+                        fluid: true,
+                        defaultOpen: true
+                    })}
+                </div>
+            </div>
+        </div>
+    )
+    .add("disabled", () =>
+        <div className="flex">
+            {createSelect({
+                disabled: true,
+                className: "mr5"
+            })}
+            {createSelect({
+                disabled: true,
+                defaultValue: "Female"
+            })}
+        </div>
+    )
+    .add("clearable", () =>
+        <div className="flex flex-column">
+            <div className="flex" style={{ marginBottom: "150px" }}>
+                {createSelect({
+                    clearable: true,
+                    defaultValue: "Female",
+                    size: "small",
+                    className: "mr5"
+                })}
+                {createSelect({
+                    clearable: true,
+                    defaultValue: "Female",
+                    className: "mr5"
+                })}
+                {createSelect({
+                    clearable: true,
+                    defaultValue: "Female",
                     size: "large"
                 })}
-                {createSearchableSelect({
-                    defaultValue: "Male",
-                    className: "mr5",
-                    size: "large"
-                })}
-                {createSearchableSelect({
-                    size: "large",
+            </div>
+            <div className="flex">
+                {createSelect({
+                    clearable: true,
+                    defaultValue: "Female",
                     defaultOpen: true
                 })}
             </div>
         </div>
     )
-    .add("disabled", () =>
-        createSearchableSelect({
-            className: "mr5",
-            disabled: true
-        })
-    )
     .add("error", () =>
         <div className="flex">
-            {createSearchableSelect({
+            {createSelect({
                 error: true,
                 className: "mr5"
             })}
-            {createSearchableSelect({
+            {createSelect({
                 error: true,
                 className: "mr5",
                 defaultOpen: true
@@ -104,51 +180,51 @@ stories()
     .add("loading", () =>
         <div className="flex flex-column">
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     size: "small",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     defaultValue: "Female",
                     size: "small",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     size: "small",
                     defaultOpen: true
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     defaultOpen: true
                 })}
             </div>
             <div className="flex">
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     size: "large",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     defaultValue: "Female",
                     size: "large",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     loading: true,
                     size: "large",
                     defaultOpen: true
@@ -156,91 +232,69 @@ stories()
             </div>
         </div>
     )
-    .add("clearable", () =>
-        <div className="flex flex-column">
-            <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
-                    clearable: true,
-                    defaultValue: "Male",
-                    size: "small",
-                    className: "mr5"
-                })}
-                {createSearchableSelect({
-                    clearable: true,
-                    defaultValue: "Male",
-                    className: "mr5"
-                })}
-                {createSearchableSelect({
-                    clearable: true,
-                    defaultValue: "Male",
-                    size: "large"
-
-                })}
-            </div>
-            <div className="flex">
-                {createSearchableSelect({
-                    clearable: true,
-                    defaultValue: "Male",
-                    defaultOpen: true
-                })}
-            </div>
+    .add("upward", () =>
+        <div style={{ marginTop: "50px" }}>
+            {createSelect({
+                upward: true,
+                defaultOpen: true
+            })}
         </div>
     )
     .add("icon", () =>
         <div className="flex flex-column">
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     size: "small",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     defaultValue: "Male",
                     size: "small",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     size: "small",
                     defaultOpen: true
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     defaultValue: "Male",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     defaultOpen: true
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     size: "large",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     defaultValue: "Male",
                     size: "large",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     size: "large",
                     defaultOpen: true
                 })}
             </div>
             <div className="flex">
-                {createSearchableSelect({
+                {createSelect({
                     icon: <MagnifierIcon />,
                     disabled: true
                 })}
@@ -250,7 +304,7 @@ stories()
     .add("item avatar", () =>
         <div className="flex flex-column">
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -258,7 +312,7 @@ stories()
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -267,7 +321,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -278,14 +332,14 @@ stories()
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
                         { ...FEMALE, avatar: { src: "https://randomuser.me/api/portraits/women/14.jpg", alt: "Female" } }
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
                         { ...FEMALE, avatar: { src: "https://randomuser.me/api/portraits/women/14.jpg", alt: "Female" } }
@@ -293,7 +347,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
                         { ...FEMALE, avatar: { src: "https://randomuser.me/api/portraits/women/14.jpg", alt: "Female" } }
@@ -303,7 +357,7 @@ stories()
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -311,7 +365,7 @@ stories()
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -320,7 +374,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, avatar: { src: "https://randomuser.me/api/portraits/men/14.jpg", alt: "Male" } },
@@ -335,7 +389,7 @@ stories()
     .add("item icons", () =>
         <div className="flex flex-column">
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -343,7 +397,7 @@ stories()
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -352,7 +406,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, icons: <CalendarIcon />, iconsPosition: "right" },
@@ -361,7 +415,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "small",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -372,14 +426,14 @@ stories()
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
                         { ...FEMALE, icons: <CalendarIcon /> }
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
                         { ...FEMALE, icons: <CalendarIcon /> }
@@ -387,7 +441,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, icons: <CalendarIcon />, iconsPosition: "right" },
                         { ...FEMALE, icons: <CalendarIcon />, iconsPosition: "right" }
@@ -395,7 +449,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
                         { ...FEMALE, icons: <CalendarIcon /> }
@@ -405,7 +459,7 @@ stories()
                 })}
             </div>
             <div className="flex" style={{ marginBottom: "150px" }}>
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -413,7 +467,7 @@ stories()
                     ],
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -422,7 +476,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, icons: <CalendarIcon />, iconsPosition: "right" },
@@ -431,7 +485,7 @@ stories()
                     defaultValue: "Female",
                     className: "mr5"
                 })}
-                {createSearchableSelect({
+                {createSelect({
                     size: "large",
                     options: [
                         { ...MALE, icons: <CalendarIcon /> },
@@ -441,5 +495,39 @@ stories()
                     className: "mr5"
                 })}
             </div>
+        </div>
+    )
+    .add("item actions", () =>
+        <div className="flex flex-column">
+            <div className="flex" style={{ marginBottom: "150px" }}>
+                {createSelect({
+                    size: "small",
+                    actions: [{ content: <a href="https://www.google.com">Google</a> }],
+                    defaultOpen: true
+                })}
+            </div>
+            <div className="flex" style={{ marginBottom: "150px" }}>
+                {createSelect({
+                    actions: [{ content: <a href="https://www.google.com">Google</a> }],
+                    defaultOpen: true
+                })}
+            </div>
+            <div className="flex">
+                {createSelect({
+                    size: "large",
+                    actions: [{ content: <a href="https://www.google.com">Google</a> }],
+                    defaultOpen: true
+                })}
+            </div>
+        </div>
+    )
+    .add("styling", () =>
+        <div className="flex">
+            {createSelect({
+                className: "border-red mr5"
+            })}
+            {createSelect({
+                style: { border: "1px solid red" }
+            })}
         </div>
     );
